Makes a functional single operation calculator
at this point after being stuck i read some code and made changes to
my code, so actually some of this are inspired (or copied) from another
person, so from now on i will disconsider my merit in this project.
I comprimise miself to build 3 different calculators in the future.